Create code
In this code, we define a function playSharpSound() that creates an AudioContext and an oscillator node using the Web Audio API. We set the oscillator type to 'sine' to generate a sharp sound. Then, we connect the oscillator to the audio destination (speakers) and start it. After 1 second, we stop the oscillator to avoid an endless sound.

Finally, we use setInterval() to schedule the playSharpSound() function to be executed every ten minutes (10 * 60 * 1000 milliseconds). This will ensure that the sharp sound is played repeatedly at the specified interval.